help-me-create-an-ai-agent/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
httpx>=0.27,<1
feedparser>=6.0,<7
google-genai>=1.0,<2
python-dotenv>=1.0,<2

